backend_foodgram/requirements.txt
asgiref==3.7.2
Django==3.2.3
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
django-filter==2.4.0
djangorestframework==3.12.4
python-dotenv==0.21.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
djoser==2.1.0
Pillow==9.2.0
reportlab==3.6.11